Initial neovim build
Neovim doesn't have a build for arm64 linux because they use GitHub
actions runners which don't include arm runners (and they don't seem to
want to maintain compiling with emulation which is totally
understandable). This docker image simply builds neovim from source, and
the architecture is just determined by the host machine.

I don't really need to update this constantly so I can just build on an
amd64 and arm64 machine when I need to for now. In the future I can test
emulation with Docker BuildKit.